- util functions moved to seperate file - response status control added - refactored based on PR comments
