+++ practice-app/app/views.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template, url_for, request
 import requests
-import json
+from .util_functions import get_game_information_from_name
 
 app = Flask(__name__)   # name is global variable returning program name which is app
 
@@ -60,29 +60,6 @@ def WorldTime():
     else:
         return render_template("worldtime.html")
 
-def get_game_id_from_name(game_name):
-    steam_api_game_list_url = "https://api.steampowered.com/ISteamApps/GetAppList/v2/"
-    response = requests.get(steam_api_game_list_url).json()
-    all_games_list = response["applist"]["apps"]
-    game_name = "".join(game_name.lower().split())
-    for game in all_games_list:
-        if game_name == "".join(game["name"].lower().split()): 
-            return game["appid"]
-    return -1
-
-def get_game_information_from_id(game_id): 
-    steam_api_get_game_url = "https://store.steampowered.com/api/appdetails?appids={game_id}"
-    response = requests.get(steam_api_get_game_url.format(game_id=game_id)).json()
-    game_information = response[str(game_id)]["data"]
-    return game_information
-
-def get_game_information_from_name(game_name): 
-    game_id = get_game_id_from_name(game_name)
-    if game_id == -1: 
-        return None
-    game_information = get_game_information_from_id(game_id)
-    return game_information
-
 @app.route("/get_game_information", methods= ["GET", "POST"])
 def get_game_information():
     error = None
@@ -93,8 +70,7 @@ def get_game_information():
         else:
             game_information = get_game_information_from_name(game_name)
             if game_information == None: 
-                error = "Game is not found!"
+                error = "Game can not be found! Try with a different name"
             else: 
-                # print(json.dumps(game_information, indent=4))
                 return render_template('show_game_information.html', response=game_information) 
     return render_template("get_game_information.html", error=error)
